Updates for planned Openwire Support
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/activemq-cpp/trunk/activemq-cpp@476372 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/util/Long.h
@@ -35,8 +35,8 @@ namespace util{
          * @param String to parse
          * @return long value
          */
-        static long parseLong( const std::string& value ){
-            long ret = 0;
+        static long long parseLong( const std::string& value ){
+            long long ret = 0;
             std::istringstream istream(value);
             istream.clear();
             istream >> ret;
@@ -48,7 +48,7 @@ namespace util{
          * @param long to convert
          * @return string representation
          */
-        static std::string toString( long value ){
+        static std::string toString( long long value ){
             std::ostringstream ostream;
             ostream << value;
             return ostream.str();
